refactor: Step2 패키지 이동 및 부둘렌주룰 의존성 추가
