+++ app/src/main/java/woowacourse/omok/model/board/PositionStatus.kt
@@ -1,0 +1,7 @@
+package woowacourse.omok.model.board
+
+enum class PositionStatus {
+    EMPTY,
+    PLACED,
+    OUT_OF_RANGE,
+}
