+++ app/src/main/java/woowacourse/omok/model/rule/RenjuFoul.kt
@@ -1,0 +1,8 @@
+package woowacourse.omok.model.rule
+
+enum class RenjuFoul {
+    THREE_BY_THREE_FOUL,
+    FOUR_BY_FOUR_FOUL,
+    OVER_FIVE_FOUL,
+    SAFE,
+}
