+++ app/src/main/java/woowacourse/omok/model/stone/Stone.kt
@@ -1,0 +1,8 @@
+package woowacourse.omok.model.stone
+
+import woowacourse.omok.model.stone.position.Position
+
+data class Stone(
+    val position: Position,
+    val stoneColor: StoneColor,
+)
