+++ app/src/main/java/woowacourse/omok/model/stone/StoneColor.kt
@@ -1,0 +1,6 @@
+package woowacourse.omok.model.stone
+
+enum class StoneColor {
+    BLACK,
+    WHITE,
+}
