+++ app/src/main/java/woowacourse/omok/model/stone/position/Position.kt
@@ -1,0 +1,6 @@
+package woowacourse.omok.model.stone.position
+
+data class Position(
+    val row: Row,
+    val col: Col,
+)
