+++ app/src/main/java/woowacourse/omok/model/stone/position/Row.kt
@@ -1,0 +1,6 @@
+package woowacourse.omok.model.stone.position
+
+@JvmInline
+value class Row(
+    val value: Int,
+)
