--- src/main/kotlin/omok/model/board/PositionStatus.kt
@@ -1,7 +1,0 @@
-package omok.model.board
-
-enum class PositionStatus {
-    EMPTY,
-    PLACED,
-    OUT_OF_RANGE,
-}
