--- src/main/kotlin/omok/model/rule/OmokReferee.kt
@@ -1,22 +1,0 @@
-package omok.model.rule
-
-import omok.model.board.Board
-import omok.model.stone.Stone
-import omok.model.stone.StoneColor
-
-class OmokReferee(
-    private val omokRule: OmokRule,
-) {
-    fun isOmok(board: Board): Boolean {
-        val lastStone: Stone = board.lastStone ?: return false
-        return omokRule.isOmok(board.stonesMap, lastStone)
-    }
-
-    fun lastStoneFoul(board: Board): RenjuFoul {
-        val lastStone: Stone = board.lastStone ?: return RenjuFoul.SAFE
-        if (lastStone.stoneColor == StoneColor.BLACK) {
-            return omokRule.checkLastBlackStoneFoul(board.stonesMap, lastStone)
-        }
-        return RenjuFoul.SAFE
-    }
-}
