--- src/main/kotlin/omok/model/stone/Stone.kt
@@ -1,8 +1,0 @@
-package omok.model.stone
-
-import omok.model.stone.position.Position
-
-data class Stone(
-    val position: Position,
-    val stoneColor: StoneColor,
-)
