--- src/main/kotlin/omok/model/stone/StoneColor.kt
@@ -1,6 +1,0 @@
-package omok.model.stone
-
-enum class StoneColor {
-    BLACK,
-    WHITE,
-}
