--- src/main/kotlin/omok/model/stone/position/Col.kt
@@ -1,6 +1,0 @@
-package omok.model.stone.position
-
-@JvmInline
-value class Col(
-    val value: Int,
-)
