--- src/main/kotlin/omok/model/stone/position/Position.kt
@@ -1,6 +1,0 @@
-package omok.model.stone.position
-
-data class Position(
-    val row: Row,
-    val col: Col,
-)
