--- src/main/kotlin/omok/model/stone/position/Row.kt
@@ -1,6 +1,0 @@
-package omok.model.stone.position
-
-@JvmInline
-value class Row(
-    val value: Int,
-)
